feat!: improve config file
BREAKING CHANGES:
Add backup_folder, dirs_file_path, ignore_file_path to config file.
Change location config.json to backup_folder/config_files/config.json
