+++ README.tr.md
@@ -53,12 +53,6 @@
 5. Başlamadan önce yapmanız gerekenler:
 
    - **example-dirs_to_backup.txt** dosyasındaki örnek **dizinleri** değiştirmelisiniz ve adını **dirs_to_backup.txt** olarak değiştirmelisiniz.
-   - `~/Documents/` dizininde `backup-for-cloud` adlı bir dizin oluşturabilirsiniz. (Eğer bilgisayarınız türkçe ise dizin adını kesinlikle değiştirmelisiniz.)
-
-     - veya yedekleme dizinini main.py bu satırdan değiştirebilirsiniz:
-       ```python
-           backup_folder: str = os.path.expanduser("~/Documents/backup-for-cloud/")
-       ```
 
 6. Script'i başlatın:
