--- config_files/backup_config.json
@@ -1,4 +1,0 @@
-{
-    "keep_count": 2,
-    "keep_enc_count": 1
-}
